refactor and bug fix

+++ utils/update_median.c
@@ -6,7 +6,7 @@
 /*   By: srosu <srosu@student.42belgium.be>        #+#  +:+       +#+         */
 /*                                               +#+#+#+#+#+   +#+            */
 /*   Created: 2026/05/11 08:47:07 by srosu            #+#    #+#              */
-/*   Updated: 2026/05/12 23:20:38 by srosu           ###   ########.fr        */
+/*   Updated: 2026/05/14 12:30:30 by srosu           ###   ########.fr        */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ void	update_median(t_stack *stack)
 			track->above_median = 0;
 		if (size > 3 && size % 2 == 1)
 		{
-			if (track->current_position == (size / 2) + 1)
+			if ((size_t) track->current_position == (size / 2) + 1)
 				track->above_median = 1;
 		}
 		track = track->next;
